Move methods about list of Builds to BuildManager
